# functions/openapi.yaml
openapi: 3.0.3
info:
  title: NovelSync Firebase Functions API
  version: 1.0.0
  description: |
    OpenAPI spec for exported HTTP Firebase Functions in this repository.

    Notes:
    - Paths are function names (for example `/generateStory`).
    - Story-protected endpoints require a Firebase ID token in `Authorization: Bearer <token>`.
    - Function `onInviteApproved` is a Firestore trigger and is intentionally excluded.
servers:
  - url: "{scheme}://{host}{base_path}"
    description: Firebase Emulator URL (configurable)
    variables:
      scheme:
        default: http
      host:
        default: 127.0.0.1:5001
      base_path:
        default: /story-6f89f/us-central1
  - url: http://127.0.0.1:5001/story-6f89f/us-central1
    description: Firebase Emulator URL (story-6f89f fixed)
  - url: https://us-central1-your-project-id.cloudfunctions.net
    description: Firebase production URL
...
security: []
paths:
...
  /searchBooks:
    get:
      tags: [Books]
      summary: Search Google Books
      operationId: searchBooks
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: maxResults
          required: false
          schema:
            type: string
            default: '10'
      responses:
        '200':
          description: Google Books API response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'